+++ src/EnginePl/oper.h
@@ -45,8 +45,8 @@
 
 
 #define Make_Oper_Key(a, t)        (((PlULong) (a) << 2) | (t))
-#define Atom_Of_Oper(k)            ((PlULong) (k) >> 2)
-#define Type_Of_Oper(k)            ((PlULong) (k) & 3)
+#define Atom_Of_Oper(k)            ((int) ((PlULong) (k) >> 2))
+#define Type_Of_Oper(k)            ((int) ((PlULong) (k) & 3))
 
 
 
@@ -60,7 +60,7 @@
 
 
 
-#define Make_Op_Mask(type)         (1<<(type))
+#define Make_Op_Mask(type)         (1 << (type))
